0. poppy

+++ 0x0E-structures_typedef/dog.h
@@ -3,7 +3,7 @@
 
 /**
  * struct dog - defines a new tpye dog
- * @age: age of dog, int
+ * @age: age of dog, float. Age in decimal point
  * @name: name of dog, char
  * @owner: owner of dog, char
  */
@@ -12,7 +12,7 @@
 struct dog
 {
 	char *name;
-	int age;
+	float age;
 	char *owner;
 }
